+++ Modules/AdminSistema/Providers/RouteServiceProvider.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace Modules\AdminSistema\Providers;
+
+use Illuminate\Foundation\Support\Providers\RouteServiceProvider as ServiceProvider;
+use Illuminate\Support\Facades\Route;
+
+class RouteServiceProvider extends ServiceProvider
+{
+    protected string $name = 'AdminSistema';
+
+    public function boot(): void
+    {
+        parent::boot();
+    }
+
+    public function map(): void
+    {
+        Route::middleware('web')->group(module_path($this->name, '/Routes/web.php'));
+    }
+}