+++ config/versiones.php
@@ -1,10 +1,11 @@
 <?php
 
 return [
-    'IEE'           => '1.22.0',
-    'BhagamApps'    => '1.21.0',
+    'IEE'           => '1.23.0',
+    'BhagamApps'    => '1.22.0',
     'User'          => '2.5.1',
     'Inventario'    => '2.15.1',
     'Apps'          => '1.5.2',
     'CrudGenerator' => '1.1.0',
+    'AdminSistema'  => '1.0.0',
 ];